+++ app/graphql/mutations/files/update_file.rb
@@ -16,7 +16,7 @@ module Mutations
       field :file, Types::File, null: true
 
       def resolve(id:, file_content:, file_name: nil)
-        file = ::FileManager::File.find(id)
+        file = ::FileRecord.find(id)
 
         if file.change_file_content!(file_content, new_filename: file_name)
           { file: }
