+++ app/graphql/mutations/files/update_file_notes.rb
@@ -13,7 +13,7 @@ module Mutations
       field :file, Types::File, null: true
 
       def resolve(id:, notes:)
-        file = ::FileManager::File.find(id)
+        file = ::FileRecord.find(id)
 
         if file.update!(notes:)
           { file: }
