+++ spec/models/file_manager/file_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe FileManager::File, type: :model do
+RSpec.describe FileRecord, type: :model do
   let(:file) { create(:file_manager_file) }
 
   it 'has a valid factory' do
